fix cursor.cursor() using argument.length
When calling cursor.cursor(), the subCursor fn calls makeCursor and its logic expects arguments.length less than 4 to trigger making subCursor.

Unfortunately subCursor calls it with 4 arguments, so we can correct this so that it will call with 3 or 4 depending on how it was called.

This is an alternate fix to @jcranendonk's since the other fix checked if value was void 0 (undefined), however it might be valid to pass undefined to these functions, so the argument.length check is probably safer.